[LOG4J2-2391] Improve ThrowableProxy performace on java 9+
Share the SecurityManager implementation of getCurrentStackTrace
with the java 9 implementation.
