+++ log4j-api-java9/src/assembly/java9.xml
@@ -35,6 +35,7 @@
         <exclude>**/Dummy.class</exclude>
         <exclude>**/spi/Provider.class</exclude>
         <exclude>**/util/PropertySource.class</exclude>
+        <exclude>**/util/PrivateSecurityManagerStackTraceUtil.class</exclude>
         <exclude>**/message/ThreadDumpMessage.class</exclude>
         <exclude>**/message/ThreadDumpMessage$ThreadInfoFactory.class</exclude>
       </excludes>
